requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
--extra-index-url https://aiinfra.pkgs.visualstudio.com/PublicPackages/_packaging/onnxruntime-cuda-12/pypi/simple/
conformer==0.3.2
absl-py==2.3.1
accelerate==1.12.0
aenum==3.1.16
aiofiles==23.2.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyascii==0.3.3
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
audioread==3.1.0
autopage==0.5.2
babel==2.17.0
backoff==2.2.1
backports-strenum==1.3.1
beautifulsoup4==4.14.3
blake3==1.0.8
bleach==6.3.0
blobfile==3.1.0
build==1.3.0
certifi>=2025.11.12
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
click-option-group==0.5.9
cliff==4.13.1
cmaes==0.12.0
cmd2==3.0.0
colored==2.3.1
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.3
conformer==0.3.2
contourpy==1.3.2
contractions==0.1.73
csvw==3.7.0
cuda-bindings==12.9.5
cuda-pathfinder==1.3.3
cuda-python==12.9.5
cycler==0.12.1
cython==3.2.3
datasets==3.1.0
debugpy==1.8.19
decorator==5.2.1
defusedxml==0.7.1
diffusers==0.36.0
dill==0.3.8
distlib==0.4.0
distro==1.9.0
dlinfo==2.0.0
einops==0.8.1
einx==0.3.0
etcd3==0.12.0
evaluate==0.4.6
exceptiongroup==1.3.1
executing==2.2.1
fastapi==0.115.4
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.20.1
flashinfer-python==0.2.5
flatbuffers==25.12.19
fonttools==4.61.1
fqdn==1.5.1
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2024.9.0
gdown==5.2.0
gradio==3.43.2
gradio-client==0.5.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h5py==3.12.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-colorlog==1.2.0
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
hyperpyyaml==1.2.2
identify==2.6.15
idna==3.11
importlib-metadata==8.7.1
importlib-resources==6.5.2
inflect==7.5.0
iniconfig==2.3.0
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
isodate==0.7.2
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter-client==8.7.0
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter-server==2.17.0
jupyter-server-terminals==0.5.3
jupyterlab==4.5.1
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.16
kaldifst==1.7.17
kiwisolver==1.4.9
language-tags==1.2.0
lark==1.3.1
lazy-loader==0.4
librosa==0.11.0
lightning==2.6.0
lightning-utilities==0.15.2
llguidance==0.7.29
llvmlite==0.46.0
loguru==0.7.3
lxml==6.0.2
mako==1.3.10
markdown==3.10
markdown-it-py==4.0.0
markupsafe==2.1.5
matcha-tts==0.0.7.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
meson==1.10.0
mistune==3.2.0
ml-dtypes==0.5.4
modelscope==1.33.0
more-itertools==10.8.0
mpi4py==4.1.1
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
narwhals==2.14.0
nbclient==0.10.4
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.13.0
nodeenv==1.10.0
notebook==7.5.1
notebook-shim==0.2.4
numba==0.63.1
numpy==1.26.4
#nvidia-cublas-cu12==12.6.4.1
#nvidia-cuda-cupti-cu12==12.6.80
#nvidia-cuda-nvrtc-cu12==12.6.77
#nvidia-cuda-runtime-cu12==12.6.77
#nvidia-cudnn-cu12==9.5.1.17
#nvidia-cufft-cu12==11.3.0.4
#nvidia-cufile-cu12==1.11.1.6
#nvidia-curand-cu12==10.3.7.77
#nvidia-cusolver-cu12==11.7.1.2
#nvidia-cusparse-cu12==12.5.4.2
#nvidia-cusparselt-cu12==0.6.3
#nvidia-ml-py==12.575.51
#nvidia-modelopt==0.33.1
#nvidia-modelopt-core==0.33.1
#nvidia-nccl-cu12==2.26.2
#nvidia-nvjitlink-cu12==12.6.85
#nvidia-nvtx-cu12==12.6.77
#nvtx==0.2.14
omegaconf==2.3.0
onnx==1.19.1
onnx-graphsurgeon==0.5.8
onnxruntime==1.23.2
onnxruntime-gpu==1.23.2
openai==2.14.0
openai-whisper==20250625
opencv-python-headless==4.11.0.86
optimum==2.1.0
optuna==2.10.1
ordered-set==4.1.0
orjson==3.11.5
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
peft==0.18.0
pexpect==4.9.0
phonemizer==3.3.0
pillow==10.3.0
platformdirs==4.5.1
pluggy==1.6.0
polygraphy==0.49.26
pooch==1.8.2
pre-commit==4.5.1
prettytable==3.17.0
prometheus-client==0.23.1
prompt-toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.0
ptyprocess==0.7.0
pulp==3.3.0
pure-eval==0.2.3
pyahocorasick==2.3.0
pyarrow==22.0.0
pycparser==2.23
pycryptodomex==3.23.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
pydub==0.25.1
pygments==2.19.2
pynvml==12.0.0
pyparsing==3.3.1
pyperclip==1.11.0
pyproject-hooks==1.2.0
pysocks==1.7.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-json-logger==4.0.0
python-multipart==0.0.21
pytorch-lightning==2.6.0
pytz==2025.2
pyworld==0.3.5
pyyaml==6.0.3
pyzmq==27.1.0
rdflib==7.5.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rich-argparse==1.7.2
rootutils==1.0.7
rpds-py==0.30.0
ruamel-yaml==0.18.17
ruamel-yaml-clib==0.2.15
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
segments==2.3.0
semantic-version==2.10.0
send2trash==1.8.3
sentencepiece==0.2.1
setuptools==79.0.1
six==1.17.0
sniffio==1.3.1
soundfile==0.13.1
soupsieve==2.8.1
soxr==1.0.0
sqlalchemy==2.0.45
stack-data==0.6.3
starlette==0.41.3
stevedore==5.6.0
strenum==0.4.15
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorrt==10.11.0.33
#tensorrt-cu12==10.11.0.33
#tensorrt-cu12-bindings==10.11.0.33
#tensorrt-cu12-libs==10.11.0.33
#tensorrt-llm==1.0.0
termcolor==3.2.0
terminado==0.18.1
textsearch==0.0.24
threadpoolctl==3.6.0
tiktoken==0.12.0
tinycss2==1.4.0
tokenizers==0.21.4
tomli==2.3.0
torch==2.7.1
torchaudio==2.7.1
torchcodec==0.9.1
torchmetrics==1.8.2
torchprofile==0.0.4
torchvision==0.22.1
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
transformers==4.53.1
#triton==3.3.1
typeguard==4.4.4
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
unidecode==1.4.0
uri-template==1.3.0
uritemplate==4.2.0
urllib3==2.6.2
uvicorn==0.40.0
virtualenv==20.35.4
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==11.0.3
werkzeug==3.1.4
wetext==0.1.2
wget==3.2
wheel==0.45.1
widgetsnbextension==4.0.15
x-transformers==2.12.0
xgrammar==0.1.21
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0

# Agent Framework Dependencies
fastmcp>=2.0.0
streamlit>=1.30.0
websocket-client>=1.9.0

# Web Hub Dependencies
flask>=3.0.0
flask-cors>=6.0.0
flask-socketio>=5.3.0
psutil>=5.9.0

# Autonomous Messaging Dependencies
apscheduler>=3.10.0

# Image Processing
pillow>=10.0.0

# Database
chromadb>=0.4.0
langchain>=0.1.0
